Refactor (rebuilt) top layer and OpenAPI layer code, enable HTTP/2 (#80)
Refactored (rebuilt) the top layer and the OpenAPI-generated layer to eliminate manually modified logic in generated code. This will make it easier to track backend updates and to add new functionality with minimal risk. Ultimately, OpenAPI-generated code should remain untouched, and all customizations should be implemented in the top layers instead. The next step is to regenerate the code for each OpenAPI service (resource) individually.

As part of the refactor, environment variables are now automatically prioritized over config variables, without requiring the "useEnvVars: true" config flag.

Enabled HTTP/2 support for the Node.js environment (powered by Undici). Undici was chosen because it provides the default built-in fetch in Node.js (since v18) and is expected to make H2 requests by default out of the box starting with Node.js v25. The default fetch remains in use for browser environments (browser fetch APIs already prioritize H2) or if undici fails to initialize. Also added the ability to pass in any custom fetch-like function when creating the client so sdk in theory can work on older versions of node.